WIP - Add table and refresh button
