+++ apps/rule-manager/client/package.json
@@ -18,6 +18,7 @@
     "@types/react-dom": "~17.0.19",
     "@vitejs/plugin-react-swc": "^3.0.0",
     "typescript": "4.5",
+    "utility-types": "^3.10.0",
     "vite": "^4.2.0"
   },
   "dependencies": {
