+++ apps/rule-manager/client/src/ts/components/pages/Rules.tsx
@@ -1,22 +1,14 @@
 import React from "react";
 import '@elastic/eui/dist/eui_theme_light.css';
-import {
-    EuiTitle,
-    EuiButton,
-    EuiPageSection,
-    EuiFlexGroup,
-    EuiFlexItem
-} from "@elastic/eui";
-import RulesTable from "../rules-table";
+import { EuiPageSection } from "@elastic/eui";
+import RulesTable from "../RulesTable";
 
 export const Rules = () => {
     return (
         <>
-
             <EuiPageSection bottomBorder={true}>
                <RulesTable />
             </EuiPageSection>
-
         </>
     );
 }
