+++ apps/rule-manager/client/src/ts/index.tsx
@@ -3,6 +3,7 @@ import * as ReactDOM from "react-dom";
 import {Page} from "./components/layout/Page";
 import "../css/reset.css";
 import "../css/typography.css";
+import "./constants/icons";
 
 // For development mode with Vite
 import 'vite/modulepreload-polyfill';
